requirements.txt
attrs==19.3.0
certifi==2019.11.28
defusedxml==0.6.0
Django==2.2.7
django-bleach==0.6.1
django-cors-headers==3.2.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.4.0
isodate==0.6.0
psycopg2==2.8.4
PyJWT==1.7.1
python-dateutil==2.8.1
requests==2.22.0
sqlparse==0.3.0
virtualenv==16.7.7
virtualenvwrapper-win==1.2.5
XlsxWriter==1.2.6
